Make `handleMessage` a `BridgeState` method
