+++ src/messages.js
@@ -1,9 +1,6 @@
 // @flow
 
-import { type PackedData, unpackData } from './data.js'
-import { closeObject, emitEvent } from './manage.js'
-import { type ChangeEvent, makeProxy, updateObjectProps } from './objects.js'
-import type { BridgeState } from './state.js'
+import { type PackedData } from './data.js'
 
 /**
  * A PackedData instance that handles object properties.
@@ -64,123 +61,4 @@ export type Message = {
   created?: Array<CreateMessage>,
   events?: Array<EventMessage>,
   returns?: Array<ReturnMessage>
-}
-
-/**
- * Handles an incoming message,
- * updating state and triggering side-effects as needed.
- */
-export function handleMessage (state: BridgeState, message: Message) {
-  // ----------------------------------------
-  // Phase 1: Get our proxies up to date.
-  // ----------------------------------------
-
-  // Handle newly-created objects:
-  if (message.created) {
-    // Pass 1: Create proxies for the new objects:
-    for (const create of message.created) {
-      state.proxies[create.localId] = makeProxy(state, create)
-    }
-
-    // Pass 2: Fill in the values:
-    for (const create of message.created) {
-      updateObjectProps(state, state.proxies[create.localId], create.props)
-    }
-  }
-
-  // Handle updated objects:
-  if (message.changed) {
-    // Pass 1: Update all the proxies:
-    let events: Array<ChangeEvent> = []
-    for (const change of message.changed) {
-      const { localId, props } = change
-      const o = state.proxies[localId]
-      if (o == null) {
-        throw new RangeError(`Invalid localId ${localId}`)
-      }
-      const newEvents = updateObjectProps(state, o, props)
-      events = events.concat(newEvents)
-    }
-
-    // Pass 2: Fire the callbacks:
-    for (const event of events) {
-      const { proxy, name, payload } = event
-      emitEvent(proxy, name, payload)
-    }
-  }
-
-  // ----------------------------------------
-  // Phase 2: Handle events & method calls
-  // ----------------------------------------
-
-  // Handle events:
-  if (message.events) {
-    for (const event of message.events) {
-      const { localId, name } = event
-      const o = localId === 0 ? state : state.proxies[localId]
-      if (o == null) continue
-      try {
-        emitEvent(o, name, unpackData(state, event, name))
-      } catch (e) {
-        emitEvent(o, 'error', e) // Payload unpacking problem
-      }
-    }
-  }
-
-  // Handle method calls:
-  if (message.calls) {
-    for (const call of message.calls) {
-      const { callId, remoteId, name } = call
-
-      try {
-        const o = state.objects[remoteId]
-        if (o == null) {
-          throw new TypeError(
-            `Cannot call method '${name}' of closed proxy (remote)`
-          )
-        }
-        if (typeof o[name] !== 'function') {
-          throw new TypeError(`'${name}' is not a function`)
-        }
-        const args = unpackData(state, call, `${name}.arguments`)
-        Promise.resolve(o[name].apply(o, args)).then(
-          value => state.emitReturn(callId, false, value),
-          e => state.emitReturn(callId, true, e)
-        )
-      } catch (e) {
-        state.emitReturn(callId, true, e)
-      }
-    }
-  }
-
-  // Handle method returns:
-  if (message.returns) {
-    for (const ret of message.returns) {
-      const { callId } = ret
-      const pendingCall = state.pendingCalls[callId]
-      if (pendingCall == null) {
-        throw new RangeError(`Invalid callId ${callId}`)
-      }
-      try {
-        pendingCall.resolve(unpackData(state, ret, '<return>'))
-      } catch (e) {
-        pendingCall.reject(e)
-      } finally {
-        delete state.pendingCalls[callId]
-      }
-    }
-  }
-
-  // ----------------------------------------
-  // Phase 3: Clean up closed objects
-  // ----------------------------------------
-
-  if (message.closed) {
-    for (const localId of message.closed) {
-      const o = state.proxies[localId]
-      if (o == null) return
-      delete state.proxies[localId]
-      closeObject(o)
-    }
-  }
 }
