+++ src/state.js
@@ -2,8 +2,9 @@
 
 import type { BridgeOptions, SendMessage, SharedClasses } from './bridge.js'
 import { Bridgeable } from './bridgeable.js'
-import { type ObjectTable, packData, packThrow } from './data.js'
+import { type ObjectTable, packData, packThrow, unpackData } from './data.js'
 import { bridgifyClass, getInstanceMagic, shareClass } from './magic.js'
+import { closeObject, emitEvent } from './manage.js'
 import type {
   CallMessage,
   ChangeMessage,
@@ -13,10 +14,13 @@ import type {
   ReturnMessage
 } from './messages.js'
 import {
+  type ChangeEvent,
   type ValueCache,
   diffObject,
   dirtyValue,
-  packObject
+  makeProxy,
+  packObject,
+  updateObjectProps
 } from './objects.js'
 
 export class BridgeState implements ObjectTable {
@@ -187,6 +191,125 @@ export class BridgeState implements ObjectTable {
     if (this.message.returns == null) this.message.returns = []
     this.message.returns.push(message)
     this.wakeup()
+  }
+
+  /**
+   * Handles an incoming message,
+   * updating state and triggering side-effects as needed.
+   */
+  handleMessage (message: Message) {
+    // ----------------------------------------
+    // Phase 1: Get our proxies up to date.
+    // ----------------------------------------
+
+    // Handle newly-created objects:
+    if (message.created) {
+      // Pass 1: Create proxies for the new objects:
+      for (const create of message.created) {
+        this.proxies[create.localId] = makeProxy(this, create)
+      }
+
+      // Pass 2: Fill in the values:
+      for (const create of message.created) {
+        updateObjectProps(this, this.proxies[create.localId], create.props)
+      }
+    }
+
+    // Handle updated objects:
+    if (message.changed) {
+      // Pass 1: Update all the proxies:
+      let events: Array<ChangeEvent> = []
+      for (const change of message.changed) {
+        const { localId, props } = change
+        const o = this.proxies[localId]
+        if (o == null) {
+          throw new RangeError(`Invalid localId ${localId}`)
+        }
+        const newEvents = updateObjectProps(this, o, props)
+        events = events.concat(newEvents)
+      }
+
+      // Pass 2: Fire the callbacks:
+      for (const event of events) {
+        const { proxy, name, payload } = event
+        emitEvent(proxy, name, payload)
+      }
+    }
+
+    // ----------------------------------------
+    // Phase 2: Handle events & method calls
+    // ----------------------------------------
+
+    // Handle events:
+    if (message.events) {
+      for (const event of message.events) {
+        const { localId, name } = event
+        const o = localId === 0 ? this : this.proxies[localId]
+        if (o == null) continue
+        try {
+          emitEvent(o, name, unpackData(this, event, name))
+        } catch (e) {
+          emitEvent(o, 'error', e) // Payload unpacking problem
+        }
+      }
+    }
+
+    // Handle method calls:
+    if (message.calls) {
+      for (const call of message.calls) {
+        const { callId, remoteId, name } = call
+
+        try {
+          const o = this.objects[remoteId]
+          if (o == null) {
+            throw new TypeError(
+              `Cannot call method '${name}' of closed proxy (remote)`
+            )
+          }
+          if (typeof o[name] !== 'function') {
+            throw new TypeError(`'${name}' is not a function`)
+          }
+          const args = unpackData(this, call, `${name}.arguments`)
+          Promise.resolve(o[name].apply(o, args)).then(
+            value => this.emitReturn(callId, false, value),
+            e => this.emitReturn(callId, true, e)
+          )
+        } catch (e) {
+          this.emitReturn(callId, true, e)
+        }
+      }
+    }
+
+    // Handle method returns:
+    if (message.returns) {
+      for (const ret of message.returns) {
+        const { callId } = ret
+        const pendingCall = this.pendingCalls[callId]
+        if (pendingCall == null) {
+          throw new RangeError(`Invalid callId ${callId}`)
+        }
+        try {
+          pendingCall.resolve(unpackData(this, ret, '<return>'))
+        } catch (e) {
+          pendingCall.reject(e)
+        } finally {
+          delete this.pendingCalls[callId]
+        }
+      }
+    }
+
+    // ----------------------------------------
+    // Phase 3: Clean up closed objects
+    // ----------------------------------------
+
+    if (message.closed) {
+      for (const localId of message.closed) {
+        const o = this.proxies[localId]
+        if (o == null) return
+        delete this.proxies[localId]
+        closeObject(o)
+      }
+    }
   }
 
   /**
